Выполнена модуляция, ДЗ 5.14
